requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
Flask-SocketIO==4.3.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2==2.8.5
python-dotenv==0.14.0
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
redis==3.5.3
simplejson==3.17.2
six==1.15.0
Werkzeug==1.0.1
